Reset import error message on deletion of an account (#419)
Fixes #381

This does not remove the import warning when deleting a specific configuration, but only when deleting the whole configuration (all the accounts attached to the konnector). It is not easy to have a "by configuration" management of the warning, as the importErrorMessage is managed on the konnector level, not an the account level.