Weboob Konnector Weboob integration was recently discussed in #496
[Cozyweboob](https://github.com/Phyks/cozyweboob) is a a wrapper script
around weboob to ease its integration in Konnector and this is the
matching connector.

If you want to give it a try yourself, you should:
* Install Weboob on your system.
* Clone [Cozyweboob](https://github.com/Phyks/cozyweboob) somewhere and
launch the `server.py` script.
* Set the weboob konnector in Konnectors.

Note that for the moment, it only handles Bill model, but extending it
to other models is just a matter of a few LoC and time, no major
troubles here.

* Add missing fields in Bill, compared to Weboob output. These fields are:
    - VAT to store the VAT part of the total amount.
    - Currency, to support bills in various currencies.
    - Extra dates fields
* Add doc on `Bill` model